+++ test/t/test_geometric_attributes.cpp
@@ -15,7 +15,7 @@
 using geom_container = std::vector<uint32_t>;
 using geom_iterator = geom_container::const_iterator;
 
-using elev_container = std::vector<int64_t>;
+using elev_container = std::vector<int32_t>;
 using elev_iterator = elev_container::const_iterator;
 
 using attr_container = std::vector<uint64_t>;
